requirements.txt
fastapi
GitPython